requirements.txt
peft>=0.10.0
trl>=0.8.0
transformers>=4.40.0
accelerate>=0.29.0
datasets>=2.19.0
torch>=2.2.0
modal>=0.62.0
numpy
pandas
pyyaml
